requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlmodel>=0.0.22
datastar-py>=0.4.0
python-multipart>=0.0.9
jinja2>=3.1.0
python-dotenv>=1.0.0
pymysql>=1.1.0
cryptography>=43.0.0
pytest>=8.0.0
httpx>=0.27.0
